Disallow creating duplicated table objects
A table object is stored as a shared_ptr in a map in the `database`
object. When it's being removed, it really is only erased from this map, which
decrements shared_ptr's `use_count()`, but doesn't necessarily
deallocate the table object, as other shared_ptr's owners can keep it
alive.
When `database` doesn't store a table in its map, but other owners keep this table alive,
another table with the same name is allowed to be created, which generally corrupts the
data. E.g. metrics data can be corrupted, because metrics inherit table's name,
so when eventually both tables (or the first/original one) is destroyed,
it causes UB, but usually ends up in throwing a "double registration"
metrics exception.
In order to avoid it, we can do the following:
1. Move table objects to a new map before they're being removed, and
   check if a table with a similar name already exists in this map when
creating a new table.
2. When removing/adding a table, clear this additional map.
3. Move checks asserting if a table (object) can be created before it's
   actually created, otherwise we'd end up with the same state.