sstables: storage: s/atomic_deleter/atomic_delete/
instead of returning a functor, let's just execute it. for better
readability.

before this change, `atomic_deleter()` returns a functor which
deletes the SSTable in batch, but this functor actually consumes
the vector of SSTable which contains the instance which holds the
the storage which in turn provides the function deleting the SSTables.
in theory, a plain call like

ssts.front()->get_storage().atomic_delete(std::move(ssts))

does not necessarily work, as C++ evalutes the expression in the order
of (in a coarse grained manner):
1. ssts.front()
2. shared_ptr::operator->()
3. SSTable::get_storage()
4. std::move(ssts)
5. storage.atomic_delete()

so, when we call the member function of `atomic_delete()` of
the derived class of storage, its owner shared_ptr is moved away
already as an element of the vector passed to this function.
we cannot assume the availability of the storage instance during
calling the hypothetical `atomic_delete()`. as it could
destroy the owner shared_ptr at anytime.

but the pattern of deleter is a little bit obscure. so in this
change, we just replace it with a plain function all. and extend
the lifecyle of the first element with a copy of the shared_ptr.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>