real_dirty_memory_accounter: document what the class is doing
The current documentation of real_dirty_memory_accounter is not
quite compelling enough.

One can find some proper documentation by digging through the
git log and reading the description of the commit which added it,
but it shouldn't be that way.

This patch replaces the current documentation of the class with something
more explanatory.

Closes #13927