cql3: add query_backend.hh
We want to move away from using query processor and storage proxy in
cql3/statements directly. So we create a new abstraction for how queries
are executed, called the query_backend. For now, this has just enough to
replace query processor in cql3/statements. To make the transition
easier, there is an accessor returning storage proxy too.
We will gradually add methods to the query backend, to replace direct
usages of storage proxy with query backend interface methods.