test.py: k expressions for pytest
Add parameter to specify pytest -k expr and pass it through.

This will allow running specific test cases instead of the whole file,
or to skip specific test(s).

NOTE: when present, the expression is passed to ALL pytest invocations,
so you should consider this if more than one pytest file will be run
when creating the expression (e.g. use 'or').

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>