db: config: Increase schema_registry_grace_period from 1s to 3min
If we have a situation when two nodes use different versions for the
current table schema, the cluster can enter a degraded state when
replica-side coordinator will keep requesting the definition of the
unknown table schema version synchronously with the request. This will
add latency to such requests as they need to wait until the schema
pull is complete. Schema pulls are known to be heavy in some cases,
and can take seconds to complete.

To reduce impact of such condition on the cluster health, we can
increase the amount of time for which temporary schema versions are
kept around and this reduce the rate of schema pulls.

There is a tradeoff to be made here. The higher the grace period, the
more memory will be used to keep older versions. But it will also
increase the likelihood of a hit in the schema registry. The period of
1s proved to be too short in practice.

Refs #4485