counters: specialize fmt::formatter<counter_{shard,cell}_view>
this is a part of a series to migrating from `operator<<(ostream&, ..)`
based formatting to fmtlib based formatting. the goal here is to enable
fmtlib to print `counter_shard_view` and `counter_cell_view` without the
help of `operator<<`.

the corresponding `operator<<()` is removed in this change, as all its
callers are now using fmtlib for formatting now.

please note, to reduce the size of the header file, unlike existing
fmt::formatter specializations, this time, instead of defining a
template function for `template<typename FormatContext>
auto format(const Type&, FormatContext&) const`, we just define a specialization
of it, namely
```
auto format(const Type&, fmt::format_context& ctx) const -> decltype(ctx.out())
```
this should reduce the size of header file, and hence shorten the
compilation time a little bit. the downside of this approach is
just we won't be able to format the customize type with {fmt} with
customized FormatContext. please note, in our use case, we don't
have such a use case -- under most of the circumstances, we use
the core function of
`fmt::format_to(OutputIt, format_string<T...>, T&&... args)` directly
or undirectly for printing, which in turn prints the args to
`fmt::buffer_context<Char>`, where `Char` is always `char`. and
{fmt} has an alias for it, namely `fmt::format_context`.

that's why we can define this specific case and it should cover
our needs without sacrificing the flexibility or functionality.

Refs #13245

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>