table: cache_update: invalidate if update failed
Invalidate the cache and clear the memtable
if _cache.update failed.

This is fine as fallback since it brings the cache into a valid
state after flushing the memtable.

In this case, just report the error and continue to invalidate
the cache and clear the memtable to free some memory
and return successfully.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>