test/network_topology_strategy_test: Test NTS with replication_factor option in test_invalid_dcs
test_invalid_dcs is a test which has a list of incorrect replication factor
values, and tries to create keyspaces with these incorrect values.

The standard way of creating a NetworkTopologyStrategy keyspace
is to specify the replication factor for each specific datacenter,
but there's also a simpler way - a user can just write: 'replication_factor': X
to convey that all of the current datacenters should have replication_factor X.

This way of creating a NetworkTopologyStrategy wasn't tested by test_invalid_dcs,
let's add it to the test to improve coverage.

Refs: https://github.com/scylladb/scylladb/issues/13986

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>