table: in update_cache, use row_cache::update_invalidating instead of row_cache::invalidate
invalidate() doesn't do some useful things -- for example it doesn't
account (for the purpose of flush control) the memory it releases
from the memtable, and doesn't make sure to clean any garbage in
the memtable's cleaner gently.

The easiest way to get these useful things is to switch from invalidate()
to update_invalidating(), which does mostly the same thing, but shares (even)
more code with update().

Technically invalidate() is slightly simpler, because it doesn't walk the
row_cache tree to find affected entries. But this shouldn't have a runtime
impact because invalidate() is only used with disabled (and thus almost
always empty) cache anyway.