feature_service::enable: don't persist features if nothing to enable
The `feature_service::enable` function will soon be called on every
topology transition command, so it's worthwhile to make it not too
expensive. Currently, it calculates the new set of enabled features and
persists them to disk, which involves a memtable flush. The function
will, most of the time, not actually enable any new features, so an
optimization is added which skips some of the work in that case.