memtable: account discarded memtable entries as freed (for flush control)
The purpose of controllers (flush controller, compaction controller) is to
keep up with background tasks with minimal resource spikes, to minimize
the worst-case impact on interactive tasks.
The goal of a controller is to keep maintainance resource usage as stable
as possible.

Therefore, controllers strive to avoid sawtooth-like patterns.
The flush controller uses a formula on the amount of total memtable memory
and amount of yet-unspooled memory. To avoid sawtooths, changes in these
memory values should be granular.

During the cache update phase, this is done by real_dirty_memory_accounter,
which is called by apply_to_incomplete for each row it frees.
However, nothing like this is done for entries which don't go through
apply_to_incomplete: for example those which aren't merged because they
exist in sstables, but not in cache, or because cache is disabled.

These entries will only be accounted as freed after the entire memtable merge
finishes. If they are a large fraction of the memtable's memory, this will
result in a sawtooth.

This patch attempts to account the memory released by discarding such entries.

Unfortunately partition_entry doesn't know its own memory usage, and the only
way to know it is to iterate over the tree. And at this point we might as well
clear it along the way.

Currently entries are discarded by putting them in mutation_cleaner.
Therefore, this patch implements the accounting by utilizing mutation_cleaner's
fallback for freed memory, and actively clearing the cleaner during the update.