test/alternator: better README.md on how to run and write tests
Improve test/alternator/README.md by adding better and more beginner-
friendly introduction to how to run the Alternator tests, as well
as a section about the philosophy of the Alternator test suite, and
some guideliness on how to write good tests in that framework.

Much of this text was copied from test/cql-pytest/README.md.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>