sstable_directory: Keep directory lister on stack
The directory_lister _lister exists as class member, but is only used
once -- when the .process() is called -- and then is closed forever.
It's simpler to keep the lister on the .process() stack.

This change also makes filesystem lister keep the copy of directory as
class member, it will be useful for the next patch as well.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>