sstable_directory: Do not collect filesystem garbage for S3-backed sstables
The sstable_directory::garbage_collect() scans /var/lib/scylla for
whatever sstable it's called for. S3-backed ones don't have anything
there, so the g.c. run is no-op. Make this call be lister virtual
method, so that only filesystem lister does this scan and the ownership
table lister becomes the real no-op. Later it will be filled with code.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>