test: move reconnect_driver to topology utils
It will be used both from `topology` and `topology_experimental_raft`.
It can't really be put in the pylib utils because that function imports
`ManagerClient` which in turn imports pylib utils, and we get a cyclical
import which python doesn't really like. The next best place to put the
`reconnect_driver` function seems to be the `topology` module.