mutation_cleaner: account freed garbage as decreasing memtable memory usage
When memtable's mutation_cleaner clears garbage versions from the memtable,
this isn't accounted (for purposes of flush control) as a decrease in memtable's
memory usage, even though it is one.

Currently, garbage versions in memtable cleaners are never cleared (they are
only cleared after the memtable's cleaner is merged into the cache's cleaner),
so this change has no effect. But in future patches they will be also cleaned
during flush, and this will become meaningful.